separate out UID mapping
